# tests/IncidentCompass.IntegrationTests/Baselines/openapi-v1.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "IncidentCompass.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://localhost/"
    }
  ],
  "paths": {
...
    "/api/v1/triage-reports": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "List compact tenant-scoped triage report summaries with stable keyset pagination.",
        "operationId": "ListTriageReports",
        "parameters": [
          {
            "name": "faultId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "serviceName",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "environment",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "classification",
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "pattern": "^-?(?:0|[1-9]\\d*)$",
              "type": [
                "integer",
                "string"
              ],
              "format": "int32"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriageReportListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Too Many Requests"
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "IncidentCompassApiKey": []
          }
        ]
      }
    },
    "/api/v1/triage-reports/{id}": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Return a triage report with backend-grounded evidence.",
        "operationId": "GetTriageReportById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriageReportDetailsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Too Many Requests"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "IncidentCompassApiKey": []
          }
        ]
      }
    },
    "/api/v1/faults/{faultId}/triage-report": {
      "get": {
        "tags": [
          "v1"
        ],
        "summary": "Return the latest triage report for a fault.",
        "operationId": "GetLatestTriageReportForFault",
        "parameters": [
          {
            "name": "faultId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriageReportDetailsResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "429": {
            "description": "Too Many Requests"
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/problem\u002Bjson": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "IncidentCompassApiKey": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "JsonElement": {},
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": [
              "null",
              "string"
            ]
          },
          "title": {
            "type": [
              "null",
              "string"
            ]
          },
          "status": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "detail": {
            "type": [
              "null",
              "string"
            ]
          },
          "instance": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
...
      "TriageReportDetailsResponse": {
        "required": [
          "id",
          "faultId",
          "status",
          "summary",
          "classification",
          "confidence",
          "documentationFit",
          "isMassIssue",
          "recommendedNextAction",
          "limitations",
          "configHash",
          "createdAtUtc",
          "supersedesReportId",
          "supersededByReportId",
          "isLatestForFault",
          "evidence"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "faultId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "classification": {
            "type": "string"
          },
          "confidence": {
            "type": "string"
          },
          "documentationFit": {
            "type": "string"
          },
          "isMassIssue": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "recommendedNextAction": {
            "type": "string"
          },
          "limitations": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "configHash": {
            "type": "string"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "supersedesReportId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "supersededByReportId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "isLatestForFault": {
            "type": "boolean"
          },
          "evidence": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TriageReportEvidenceResponse"
            }
          }
        }
      },
      "TriageReportEvidenceResponse": {
        "required": [
          "id",
          "kind",
          "artifactId",
          "reference",
          "quote",
          "score",
          "artifactKind",
          "artifactDomainRef",
          "artifactPayload"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "kind": {
            "type": "string"
          },
          "artifactId": {
            "type": "string",
            "format": "uuid"
          },
          "reference": {
            "type": "string"
          },
          "quote": {
            "type": [
              "null",
              "string"
            ]
          },
          "score": {
            "pattern": "^-?(?:0|[1-9]\\d*)(?:\\.\\d\u002B)?(?:[eE][\u002B-]?\\d\u002B)?$",
            "type": [
              "null",
              "number",
              "string"
            ],
            "format": "double"
          },
          "artifactKind": {
            "type": "string"
          },
          "artifactDomainRef": {
            "type": [
              "null",
              "string"
            ]
          },
          "artifactPayload": {
            "$ref": "#/components/schemas/JsonElement"
          }
        }
      },
      "TriageReportListItemResponse": {
        "required": [
          "id",
          "faultId",
          "status",
          "summary",
          "classification",
          "confidence",
          "isMassIssue",
          "recommendedNextAction",
          "createdAtUtc",
          "serviceName",
          "environment",
          "supersedesReportId",
          "supersededByReportId",
          "isLatestForFault"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "faultId": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string"
          },
          "summary": {
            "type": "string"
          },
          "classification": {
            "type": "string"
          },
          "confidence": {
            "type": "string"
          },
          "isMassIssue": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "recommendedNextAction": {
            "type": "string"
          },
          "createdAtUtc": {
            "type": "string",
            "format": "date-time"
          },
          "serviceName": {
            "type": "string"
          },
          "environment": {
            "type": "string"
          },
          "supersedesReportId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "supersededByReportId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "isLatestForFault": {
            "type": "boolean"
          }
        }
      },
      "TriageReportListResponse": {
        "required": [
          "reports",
          "nextCursor"
        ],
        "type": "object",
        "properties": {
          "reports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TriageReportListItemResponse"
            }
          },
          "nextCursor": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    },
    "securitySchemes": {
      "IncidentCompassApiKey": {
        "type": "apiKey",
        "description": "Host-issued IncidentCompass API key.",
        "name": "X-IncidentCompass-Key",
        "in": "header"
      }
    }
  },
...
}